pyproject.toml
[tool.poetry]
name = "bridging-program-qa-automation"
version = "0.1.0"
description = ""
authors = ["Sonialyc <132089675+Sonialyc@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.12.0"
pytest = "^7.4.2"
junit-xml = "^1.9"
pytest-html = "^4.0.2"
pillow = "^10.0.1"
pyautoit = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
